Don't send file header to server
This should improve parallelizability of uploading files, as well as
improving the privacy of uploaders. Big ass refactor.

I'm trying to move as many implementation details, such as using blake3
for hasing or rkyv for serialization, to the bfsp library as possible.
Hopefully the next commit will be refactoring the code to be more readable.

Also added a test for file header consistency.