Proper encryption 🎉
*insert Charlie celebrating meme*
THE CLIENT NOW SUPPORTS END TO END ENCRYPTION!!!!!!!!!!! Lots of work to
do on the client side to turn bfsp into a proper library for
implementing a client anywhere, but this is great process. We use
xchacha20poly1035 for encryption, because it has the nice property of
being very fast even on platforms that don't have native
implementations (web, mobile, weird architectures, etc.)

I'm trying to keep as many implementaiton details, such as hash or
encryption algorithm being used on the server as possible, so we can
make updates to the protocol easily. Next steps should be refactoring,
adding compression, and turing the CLI into an actual cli, not just a
"sync tux_huge.png" program.