Allow deleting chunks + optimized insert_chunk
Firstly, we now have the ability to delete chunks uploaded, which is
useful for uploading files that got modified, as well as actually
deleting files.

Also implemented a small (but borked) optimization where if a user is
uploading a chunk with a hash that's already found on the server, we
just don't do the rest of the work of copying into the chunk. This
doesn't work when there's more than one user on the server, since that
means multiple users can't own a chunk. Easy fix at some point.