switch file metadata id to string, change file_metadata primary key
this is to make using the file metadata as a uuid a bit easier (plus a
bit easier to read).

We also changed the primary key to be a composite one (a combination of
the file id and the user id). That way, one user can have the same file
id as another without conflict. Nice!