Added storage caps for users
Having limits on the amount of data that users can store is obviously
useful so users can pay for what they use up to a certain point,
restricting abusive users, etc. By default we don't have a limit unless
a user is given a specific one, which is done by adding it to a table.
Once we get some more billing work done, we could have big_central hit
the internal API

We also started counting file_metadata in chunk usage. The reason for
this is to prevent abuse. Any user can upload any data into
file_metadata, so ensuring that we check limits there is ideal. Once we
change chunk_metadata to encrypt some of the data, we should also count
that. In reality, this is a KB at *most* per file, so this isn't an
issue for non-abusive users